refactor: Migrate to TypeScript init
