+++ server.ts
@@ -1,6 +1,5 @@
-// server.js
-const app = require("./src/app");
-const config = require("./src/config/config");
+import app from "./src/app";
+import config from "./src/config/config";
 
 const server = app.listen(config.port, () => {
   console.log(`Server listening on port ${config.port}`);
