+++ src/config/config.ts
@@ -1,8 +1,9 @@
-// src/config/config.js (예시)
-require("dotenv").config();
+import "dotenv/config";
 
-module.exports = {
+const config = {
   port: process.env.PORT || 5050,
   jwtSecret: process.env.JWT_SECRET || "your-secret-key",
   jwtExpiration: process.env.JWT_EXPIRATION || "1h",
 };
+
+export default config;
